+++ R/defaults.R
@@ -111,6 +111,7 @@ opts_chunk_attr = local({
   opts$fig.dim = 'list'
   opts$R.options = 'list'
   opts$cache.comments = 'logical'
+  opts$animation.hook = list('ffmpeg', 'gifski')
   opts
 })
 
@@ -158,8 +159,7 @@ set_alias = function(...) {
 opts_knit = new_defaults(list(
   progress = TRUE, verbose = FALSE, width = 75L, eval.after = 'fig.cap',
   base.dir = NULL, base.url = NULL, root.dir = NULL, child.path = '',
-  upload.fun = identity, animation.fun = hook_ffmpeg_html,
-  global.device = FALSE, global.par = FALSE,
+  upload.fun = identity, global.device = FALSE, global.par = FALSE,
   concordance = FALSE, documentation = 1L, self.contained = TRUE,
   unnamed.chunk.label = 'unnamed-chunk', highr.opts = NULL,